requirement.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.1
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.8
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
requests==2.32.3
resend==2.4.0
rest-framework-simplejwt==0.0.2
sendgrid==6.11.0
sib-api-v3-sdk==7.6.0
six==1.16.0
sqlparse==0.5.1
starkbank-ecdsa==2.2.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
